requirements.txt
strands-agents
bedrock-agentcore
snowflake-connector-python
python-dotenv
snowflake-core
snowflake-snowpark-python
fastapi